Fix the datatype error in fake op

+++ tensorrt_llm/_torch/auto_deploy/custom_ops/triton_backend_mamba.py
@@ -176,7 +176,7 @@ def _triton_cached_ssm_transform_fake(
     return torch.empty_like(
         hidden_states,
         memory_format=torch.contiguous_format,
-        dtype=torch.float32,
+        dtype=hidden_states.dtype,
     )
